extracted addProjectsPage nested class to diff files
