+++ src/AddProjectsPage/AddProjectsPage.tsx
@@ -14,6 +14,12 @@ import { getTags } from '../actions/tagsActions';
 import { getCategories } from '../actions/categoryActions';
 import { Store, AddProjectProps, Action } from '../types/Redux';
 import { Redirect } from 'react-router';
+import StatusOptionsComponent from './StatusOptionsComponent';
+import ChosenTeam from './ChosenTeam';
+import ChosenTags from './ChosenTags';
+import TagOptionsComponent from './TagOptionsComponent';
+import CategoriesOptionsComponent from './CategoriesOptionsComponent';
+import TeamOptionsComponent from './TeamOptionsComponent';
 
 class AddProjectsPage extends React.Component<
   AddProjectProps,
@@ -342,253 +348,6 @@ class AddProjectsPage extends React.Component<
           to={'/projects/' + this.state.projIdRedirect}
         />
       );
-    }
-    class StatusOptionsComponent extends React.Component<{
-      onFormChange: any;
-    }> {
-      render() {
-        let statusOptionsArray = ['Active', 'Completed'];
-
-        var statusComponent = statusOptionsArray.map(
-          (status: string, index: number) => {
-            return (
-              <input
-                key={'status_' + index}
-                type="button"
-                name="status"
-                value={status}
-                onClick={this.props.onFormChange}
-                className="new-project-dropdown-text"
-              />
-            );
-          }
-        );
-
-        return statusComponent;
-      }
-    }
-    class TeamOptionsComponent extends React.Component<{
-      allUsers: any;
-      user: any;
-      onFormChange: any;
-      teamFilter: any;
-    }> {
-      render() {
-        let teamOptionsComponent: JSX.Element[];
-        let usersFromStore = this.props.allUsers!;
-        let username = this.props.user.username;
-        if (usersFromStore instanceof Array) {
-          usersFromStore = usersFromStore.filter(
-            user => user.username !== username
-          );
-          teamOptionsComponent = usersFromStore.map(
-            (users: any, index: number) => {
-              return (
-                <input
-                  key={'users_' + index}
-                  type="button"
-                  name="team"
-                  value={users.username}
-                  onClick={this.props.onFormChange}
-                  className="new-project-dropdown-text"
-                />
-              );
-            }
-          );
-        }
-        return (
-          <div>
-            <input
-              className="search-input-box"
-              type="text"
-              placeholder="Search for Teammates"
-              id="teamSearch"
-              onKeyUp={this.props.teamFilter}
-            />
-            {teamOptionsComponent!}
-          </div>
-        );
-      }
-    }
-
-    class CategoriesOptionsComponent extends React.Component<{
-      categories: any;
-      onFormChange: any;
-      categoryFilter: any;
-    }> {
-      render() {
-        let categoryOptionsComponent: JSX.Element[];
-        let categoriesFromStore = this.props.categories!;
-        if (categoriesFromStore instanceof Array) {
-          categoryOptionsComponent = categoriesFromStore.map(
-            (category: any, index: number) => {
-              return (
-                <input
-                  key={'categories_' + index}
-                  type="button"
-                  name="category"
-                  value={category.categoryName}
-                  onClick={this.props.onFormChange}
-                  className="new-project-dropdown-text"
-                />
-              );
-            }
-          );
-        }
-        return (
-          <div>
-            <input
-              className="search-input-box"
-              type="text"
-              placeholder="Search Categories"
-              id="categorySearch"
-              onKeyUp={this.props.categoryFilter}
-            />
-            {categoryOptionsComponent!}
-          </div>
-        );
-      }
-    }
-
-    class TagOptionsComponent extends React.Component<{
-      formChange: any;
-      tags: any;
-      tagFilter: any;
-    }> {
-      render() {
-        let tagOptionsComponent: JSX.Element[];
-        let tagsFromStore = this.props.tags!;
-        if (tagsFromStore instanceof Array) {
-          tagOptionsComponent = tagsFromStore.map(
-            (tagObject: any, index: number) => {
-              return (
-                <input
-                  key={'tags_' + index}
-                  type="button"
-                  name="tags"
-                  value={tagObject.tagName}
-                  onClick={this.props.formChange}
-                  className="new-project-dropdown-text"
-                />
-              );
-            }
-          );
-        }
-        return (
-          <div>
-            <input
-              className="search-input-box"
-              type="text"
-              placeholder="Search / Add Tags"
-              id="tagSearch"
-              onKeyUp={this.props.tagFilter}
-            />
-            {tagOptionsComponent!}
-          </div>
-        );
-      }
-    }
-
-    class ChosenTags extends React.Component<{
-      tags: any;
-      handleOptionRemoval: any;
-    }> {
-      render() {
-        var chosenTags;
-        if (!this.props.tags) {
-          return null;
-        }
-
-        let tags = this.props.tags.slice();
-
-        if (tags.length === 0) {
-          chosenTags = null;
-        } else {
-          chosenTags = tags.map((tagName: string, index: number) => {
-            return (
-              <div className="tag-container" key={index}>
-                <input
-                  type="button"
-                  className="new-project-chosen-tag"
-                  value={tagName}
-                />
-                <button
-                  type="button"
-                  className="remove-tag-btn"
-                  onClick={e => this.props.handleOptionRemoval(e, 'tags', tags)}
-                >
-                  X
-                </button>
-              </div>
-            );
-          });
-        }
-        return <div className="array-of-tags">{chosenTags}</div>;
-      }
-    }
-
-    class ChosenTeam extends React.Component<{
-      team: any;
-      handleOptionRemoval: any;
-    }> {
-      render() {
-        let chosenTeam;
-        let team = Object.assign([], this.props.team);
-        if (team.length === 0) {
-          chosenTeam = null;
-        } else {
-          if (team.length === 1) {
-            chosenTeam = (
-              <div className="tag-container" key={1}>
-                <input
-                  type="button"
-                  className="new-project-chosen-tag"
-                  value={team}
-                />
-                <button
-                  type="button"
-                  className="remove-tag-btn"
-                  onClick={e =>
-                    this.props.handleOptionRemoval(
-                      e,
-                      'team',
-                      Object.assign([], this.props.team)
-                    )
-                  }
-                >
-                  X
-                </button>
-              </div>
-            );
-          } else {
-            chosenTeam = team.map((teamMemeber: string, index: number) => {
-              return (
-                <div className="tag-container" key={index}>
-                  <input
-                    type="button"
-                    className="new-project-chosen-tag"
-                    value={teamMemeber}
-                  />
-                  <button
-                    type="button"
-                    className="remove-tag-btn"
-                    onClick={e =>
-                      this.props.handleOptionRemoval(
-                        e,
-                        'team',
-                        Object.assign([], this.props.team)
-                      )
-                    }
-                  >
-                    X
-                  </button>
-                </div>
-              );
-            });
-          }
-        }
-        return chosenTeam;
-      }
     }
 
     return (
